Changed many data types from "long" to "off_t" to allow the function "bu_fseek" to work correctly on Windows 64bit.
svn:revision:54729
svn:branch:trunk
svn:account:r_weiss

+++ src/fbed/char.c
@@ -77,8 +77,8 @@ do_line(int xpos, int ypos, const char* line, RGBpixel (*menu_border))
 	     == EOF
 	    )
 	{
-	    fb_log( "fseek() to %ld failed.\n",
-		    (long)(SWABV(font.dir[char_id].addr) + font.offset)
+	    fb_log( "fseek() to %zd failed.\n",
+		    (SWABV(font.dir[char_id].addr) + font.offset)
 		);
 	    return;
 	}
