+++ src/util/pixscale.c
@@ -45,11 +45,11 @@
 
 unsigned char *outbuf;
 unsigned char *buffer;
-int scanlen;			/* length of infile (and buffer) scanlines */
-int buflines;			/* Number of lines held in buffer */
-int buf_start = -1000;		/* First line in buffer */
+ssize_t scanlen;		/* length of infile (and buffer) scanlines */
+ssize_t buflines;		/* Number of lines held in buffer */
+off_t buf_start = -1000;	/* First line in buffer */
 
-int bufy;				/* y coordinate in buffer */
+ssize_t bufy;				/* y coordinate in buffer */
 FILE *buffp;
 static char *file_name;
 
@@ -79,7 +79,7 @@ Usage: pixscale [-h] [-r] [-s squareinsize] [-w inwidth] [-n inheight]\n\
 void
 fill_buffer(int y)
 {
-    static int file_pos = 0;
+    static off_t file_pos = 0;
     size_t ret;
 
     buf_start = y - buflines/2;
